Add support for `update_only` option to nested attributes of collections
Now, if a `accepts_nested_attributes_for` is defined for a collection association,
we can specify `update_only: true`. Then, new records will not be created in the
collection and will simply be ignored.

Before this commit, update_only option was only supported by one-to-one associations.
For collection associations it was simply ignored.