Remove ruby version constraint
Having an `.rvmrc` file infers that the contributor has RVM isntalled,
while this is not always the case, due to an number of alternatives.

Not having standard `.ruby-version`, `.ruby-gemset`, and `.version.conf`
in `.gitignore` is also putting a roadblock for those using alternative
ruby version managers and modern approach to version management
configuration.

Those files should remain in local repositories and not be tracked.